(#73) AssemblyContext: move methods from the Assemblies static class
