+++ Cesium.CodeGen.Tests/CodeGenTestBase.cs
@@ -1,7 +1,6 @@
 using System.Text;
 using Cesium.CodeGen.Contexts;
 using Cesium.CodeGen.Extensions;
-using Cesium.CodeGen.Generators;
 using Cesium.Parser;
 using Cesium.Test.Framework;
 using Mono.Cecil;
@@ -21,7 +20,7 @@ public abstract class CodeGenTestBase : VerifyTestBase
     }
 
     private static AssemblyContext CreateAssembly(TargetRuntimeDescriptor? targetRuntime) =>
-        Assemblies.Create(
+        AssemblyContext.Create(
             new AssemblyNameDefinition("test", new Version()),
             ModuleKind.Console,
             targetRuntime,
@@ -42,7 +41,7 @@ public abstract class CodeGenTestBase : VerifyTestBase
             if (parser.TokenStream.Peek().Kind != CTokenType.End)
                 throw new Exception($"Excessive output after the end of a translation unit at {lexer.Position}.");
 
-            Assemblies.EmitTranslationUnit(context, translationUnit.Ok.Value.ToIntermediate());
+            context.EmitTranslationUnit(translationUnit.Ok.Value.ToIntermediate());
         }
     }
