+++ Cesium.CodeGen/Ir/TopLevel/ITopLevelDeclaration.cs
@@ -4,5 +4,5 @@ namespace Cesium.CodeGen.Ir.TopLevel;
 
 public interface ITopLevelNode
 {
-    public void Emit(TranslationUnitContext context);
+    public void EmitTo(TranslationUnitContext context);
 }
